# cmd/server/openapi.yaml
openapi: 3.1.0
info:
  title: SubShare Admin API
  version: 1.0.0
  description: |
    Versioned administration API for SubShare. Browser sessions use the
    `subshare_admin_session` cookie and CSRF protection on unsafe requests. API
    clients use a scoped bearer token created in Security settings.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Search and paginate users
      parameters:
        - { $ref: "#/components/parameters/Query" }
        - { name: status, in: query, schema: { type: string, enum: [all, active, expired, paused, blocked, limited] } }
        - { name: sort, in: query, schema: { type: string, enum: [created_desc, name_asc, expires_asc] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: User page
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/PageEnvelope" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/UserSummary" }
        default: { $ref: "#/components/responses/Error" }
...
  /keys:
    get:
      tags: [Keys]
      summary: Search and paginate keys
      parameters:
        - { $ref: "#/components/parameters/Query" }
        - { name: status, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: Key page
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/PageEnvelope" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/KeySummary" }
        default: { $ref: "#/components/responses/Error" }
  /sources:
    get:
      tags: [Operations]
      summary: Paginate safe external subscription source summaries
      parameters:
        - { $ref: "#/components/parameters/Query" }
        - { name: status, in: query, schema: { type: string, enum: [all, disabled, idle, syncing, ok, error] } }
        - { name: sort, in: query, schema: { type: string, enum: [last_sync_desc, name_asc, created_asc] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: Source page without full URLs or HWID values
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/PageEnvelope" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/SourceSummary" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [Operations]
      summary: Import a validated external source
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceInput" }
      responses:
        "201": { description: Source and imported keys created atomically }
        "409": { $ref: "#/components/responses/Error" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: subshare_admin_session
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: ss_<random>
  parameters:
...
    Query:
      name: query
      in: query
      schema: { type: string, maxLength: 200 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        field_errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        request_id: { type: string }
    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta: { $ref: "#/components/schemas/PageMeta" }
...
    UserSummary:
      type: object
      required: [id, name, status, effective_status]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        status: { type: string }
        effective_status: { type: string, enum: [active, expired, paused, blocked, limited] }
        starts_at: { type: string }
        expires_at: { type: string }
        max_devices: { type: integer }
        connected_device_count: { type: integer }
        created_at: { type: string, format: date-time }
    KeySummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        label: { type: string }
        category: { type: string }
        kind: { type: string }
        status: { type: string }
        check_status: { type: string }
        check_error: { type: string }
        last_latency_ms: { type: integer, format: int64 }
        last_checked_at: { type: string }
        external_source_id: { type: integer, format: int64 }
        external_source_name: { type: string }
        created_at: { type: string, format: date-time }
    SourceSummary:
      type: object
      required: [id, name, source_url_masked, enabled, has_hwid_value, import_status, imported_keys]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
        key_category: { type: string }
        source_url_masked: { type: string }
        enabled: { type: boolean }
        pass_hwid: { type: boolean }
        has_hwid_value: { type: boolean }
        last_import_count: { type: integer }
        imported_keys: { type: integer }
        import_status: { type: string, enum: [idle, syncing, ok, error] }
        last_error: { type: string }
        last_synced_at: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
...
    SourceInput:
      type: object
      required: [name, category, key_category, key_insert_mode, source_url, enabled, pass_hwid]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: Trimmed source name, limited to 64 Unicode code points.
        category: { type: string }
        key_category: { type: string }
        key_insert_mode: { type: string, enum: [top, bottom] }
        source_url: { type: string, format: uri }
        enabled: { type: boolean }
        apply_remote_metadata: { type: boolean }
        pass_hwid: { type: boolean }
        hwid_version: { type: string }
        hwid_model_name: { type: string }
        hwid_value:
          type: string
          description: Omit on update to preserve the stored secret.
        clear_hwid_value:
          type: boolean
          description: Explicitly remove the stored HWID value.
        raw_body: { type: string, maxLength: 1048576 }
        raw_content_type: { type: string }
        raw_final_url: { type: string, format: uri }
...